requirements.txt
fastapi
uvicorn[standard]
python-multipart
llama-cpp-python
openai
diskcache
python-dotenv
